Changed the XML string precision from 12 to 16, and added a bugfix to lib/vdf/Copy2VDF.cpp.  Multiple timesteps in the CAM data conversion are now successful.

+++ lib/vdf/Copy2VDF.cpp
@@ -77,7 +77,8 @@ void Copy2VDF::GetTimeMap(
 	//
 	// Get user times from .vdf
 	//
-	vector <float> ncdftimes;
+	//vector <float> ncdftimes;
+	vector <double> ncdftimes;
 	for (int i=0; i<DCData->GetNumTimeSteps(); i++) { 
 		ncdftimes.push_back(DCData->GetTSUserTime(i));
 	}
@@ -85,7 +86,8 @@ void Copy2VDF::GetTimeMap(
 	//
 	// Get user times from netCDF files
 	//
-	vector <float> vdctimes;
+	//vector <float> vdctimes;
+	vector <double> vdctimes;
 	for (int i=0; i<vdfio->GetNumTimeSteps(); i++) { 
 		vdctimes.push_back(vdfio->GetTSUserTime(i));
 	}
@@ -97,11 +99,11 @@ void Copy2VDF::GetTimeMap(
 	//
 	// Cross-reference ncdf and vdc times
 	//
-	float ncdftime; 
+	double ncdftime; 
 	for (int i=startts; i<numts+startts && i<DCData->GetNumTimeSteps(); i++) {
 		ncdftime = ncdftimes[i];
 
-		vector <float>::iterator itr = find(
+		vector <double>::iterator itr = find(
 			vdctimes.begin(), vdctimes.end(), ncdftime
 		);
